% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getCountryYearCom}
\alias{getCountryYearCom}
\title{Display all combinations of country and year for DHS surveys.}
\usage{
getCountryYearCom(conn)
}
\arguments{
\item{conn}{a Formal class MySQLConnection established with DBI and RMySQL}
}
\value{
a data frame will the unique combinations of country and year for DHS surveys.
used as a helper for running functions like `get_tMax()` and `getPrecip()`.
}
\description{
\code{getCountryYearCom()} will display all combinations of country and year for DHS surveys
}
\examples{
showCountryYearCom(conn = mydb)


}
