% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPrecip}
\alias{getPrecip}
\title{Get CHIRPS data for a specified country and time period}
\usage{
getPrecip(conn, countryCode, Year, timeFrame)
}
\arguments{
\item{conn}{a Formal class MySQLConnection established with `RMySQL`}

\item{countryCode}{a 3 digit character vector representing the country where spatial points should be extracted from. This is the
'COUNTRY' column in the countries table in the jgnutritionsecurity database. Run \code{\link{getCountryYearCom}} to find unique country and year combinations.}

\item{Year}{a 4 digit character vector representing the Year the survey was conducted. This variable is called 'DHSYEAR' in DHS data.}

\item{timeFrame}{numeric vector representing the number of years of precipitation data required.
\itemize{
\item\strong{timeFrame = 0} will return data from the DHS survey year
\item\strong{timeFrame = 1} will return data from the DHS survey year and 1 year prior
\item etc...
}}
}
\value{
A data frame with monthly precipitation values in mm/month for each spatial observation. Additional attributes for spatial observations will be preserved.
See CHIRPS website for more information. \url{https://www.chc.ucsb.edu/data/chirps}
}
\description{
\code{getPrecip()} takes DHS spatial point data from the jgnutritionsecurity database and reads geotiff data from an HTTP web protocol to calculate monthly precipitation values at each point of interest.
}
\examples{
getPrecip(conn = mydb,
 countryCode = "024",
  Year = "2015",
   timeFrame = 4)

Will return precipitation data for all DHS points in
Angola from 01-01-2011 to 12-31-2015

}
