% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uniqueID}
\alias{uniqueID}
\title{Find unique identifier for a specified dataframe}
\usage{
uniqueID(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
The number of unique variables by column. The unique ID should equal the total number of observations in
         the dataframe.
}
\description{
\code{uniqueID()} When used with lapply on a data frame it counts the number of unique variables in each column
}
\examples{
lapply(dataframe, FUN = uniqueID)

}
