% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPrecip}
\alias{getPrecip}
\title{Get CHIRPS data for a specified country and time period}
\usage{
getPrecip(conn, countryCode, Year, startDate, endDate)
}
\arguments{
\item{conn}{a Formal class MySQLConnection established with `RMySQL`}

\item{countryCode}{a 3 digit character vector representing the country where spatial points should be extracted from. This is the
'COUNTRY' column in the countries table in the jgnutritionsecurity database. Run \code{\link{getCountryYearCom}} to find unique country and year combinations.}

\item{Year}{a 4 digit character vector representing the Year the survey was conducted.}

\item{startDate, endDate}{YYYY-MM-DD for start date and end date.}
}
\value{
A data frame with monthly precipitation values in mm/month for each spatial observation. Additional attributes for spatial observations will be preserved.
See CHIRPS website for more information. \url{https://www.chc.ucsb.edu/data/chirps}
}
\description{
\code{getPrecip()} takes DHS spatial point data from the jgnutritionsecurity database and reads geotiff data from an HTTP web protocol to calculate monthly precipitation values at each point of interest.
}
\examples{
getPrecip(conn = mydb,
 countryCode = "024",
  Year = "2015",
   startDate = "2014-01-01",
   endDate = "2016-12-31")

Will return precipitation data for all DHS points in
Angola from 01-01-2014 to 12-31-2016
}
